+++ telecode.sql
@@ -1,7 +1,7 @@
 DROP TABLE IF EXISTS configs CASCADE;
 CREATE TABLE configs (
   id serial PRIMARY KEY,
-  chatId int not NULL,
+  chatId bigint not NULL,
   inputMode VARCHAR(32) NOT NULL DEFAULT '',
   theme varchar(32) NOT NULL
 );
